Transcendence02/backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
django-otp
django-two-factor-auth
psycopg2-binary
gunicorn
phonenumbers
django-cors-headers
django-sslserver
pyotp
Pillow
django-csp
social-auth-app-django